+++ lib/aws_auth/query_parameters.ex
@@ -35,28 +35,28 @@ defmodule AWSAuth.QueryParameters do
     params =
       case uri.query do
         nil ->
-          Map.new()
+          []
 
         _ ->
-          URI.decode_query(uri.query)
+          Enum.to_list(URI.query_decoder(uri.query))
       end
 
     # Default expiration is 15 minutes (900 seconds), max is 7 days (604800 seconds)
-    expires_in = Keyword.get(opts, :expires_in, 900)
+    expires_in = opts |> Keyword.get(:expires_in, 900) |> normalize_expires_in!()
     uri_escape_path = Keyword.get(opts, :uri_escape_path, true)
 
     params =
       params
-      |> Map.put("X-Amz-Algorithm", "AWS4-HMAC-SHA256")
-      |> Map.put("X-Amz-Credential", "#{access_key}/#{scope}")
-      |> Map.put("X-Amz-Date", amz_date)
-      |> Map.put("X-Amz-Expires", to_string(expires_in))
-      |> Map.put("X-Amz-SignedHeaders", "#{Map.keys(headers) |> Enum.join(";")}")
+      |> put_param("X-Amz-Algorithm", "AWS4-HMAC-SHA256")
+      |> put_param("X-Amz-Credential", "#{access_key}/#{scope}")
+      |> put_param("X-Amz-Date", amz_date)
+      |> put_param("X-Amz-Expires", to_string(expires_in))
+      |> put_param("X-Amz-SignedHeaders", AWSAuth.Utils.signed_headers(headers))
 
     # Add session token to query params if provided (for temporary credentials)
     params =
       if session_token do
-        Map.put(params, "X-Amz-Security-Token", session_token)
+        put_param(params, "X-Amz-Security-Token", session_token)
       else
         params
       end
@@ -69,7 +69,7 @@ defmodule AWSAuth.QueryParameters do
     string_to_sign =
       AWSAuth.Utils.build_canonical_request(
         http_method,
-        uri.path,
+        uri.path || "/",
         params,
         headers,
         hashed_payload,
@@ -81,9 +81,30 @@ defmodule AWSAuth.QueryParameters do
       AWSAuth.Utils.build_signing_key(secret_key, date, region, service)
       |> AWSAuth.Utils.build_signature(string_to_sign)
 
-    params = params |> Map.put("X-Amz-Signature", signature)
-    query_string = URI.encode_query(params) |> String.replace("+", "%20")
+    params = put_param(params, "X-Amz-Signature", signature)
+    query_string = AWSAuth.Utils.canonical_query_string(params)
 
     "#{uri.scheme}://#{uri.authority}#{uri.path || "/"}?#{query_string}"
+  end
+
+  defp put_param(params, key, value) do
+    [{key, value} | Enum.reject(params, fn {existing_key, _value} -> existing_key == key end)]
+  end
+
+  defp normalize_expires_in!(expires_in)
+       when is_integer(expires_in) and expires_in >= 1 and expires_in <= 604_800, do: expires_in
+
+  defp normalize_expires_in!(expires_in) when is_binary(expires_in) do
+    case Integer.parse(expires_in) do
+      {value, ""} -> normalize_expires_in!(value)
+      _other -> invalid_expires_in!(expires_in)
+    end
+  end
+
+  defp normalize_expires_in!(expires_in), do: invalid_expires_in!(expires_in)
+
+  defp invalid_expires_in!(expires_in) do
+    raise ArgumentError,
+          ":expires_in must be an integer between 1 and 604800 seconds, got: #{inspect(expires_in)}"
   end
 end